hr_system_asmald/requirements.txt
aiogram==2.25.1
aiohttp==3.8.4
aiosignal==1.3.1
anyio==3.7.0
asgiref==3.7.2
async-timeout==4.0.2
asyncpg==0.27.0
attrs==22.2.0
Babel==2.9.1
certifi==2022.12.7
charset-normalizer==3.1.0
distlib==0.3.6
Django==4.2.2
django-admin-datta==1.0.10
django-birthday==0.1.4
django-crispy-forms==2.0
django-cron==0.6.0
environs==8.0.0
eskiz-sms==0.2.3
et-xmlfile==1.1.0
filelock==3.10.0
frozenlist==1.3.3
gunicorn==20.1.0
h11==0.14.0
httpcore==0.16.3
httpx==0.23.3
idna==3.4
magic-filter==1.0.9
marshmallow==3.19.0
multidict==6.0.4
numpy==1.24.2
openpyxl==3.1.2
packaging==23.0
pandas==2.0.0
Pillow==9.5.0
platformdirs==3.1.1
psycopg2-binary==2.9.6
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
requests==2.31.0
rfc3986==1.5.0
six==1.16.0
sniffio==1.3.0
sqlparse==0.4.4
tzdata==2023.3
urllib3==2.0.3
virtualenv==20.21.0
whitenoise==6.5.0
xlwt==1.3.0
yarl==1.8.2
